tests: Fix rhbz690819.sh to skip when attach-method is libvirt.
This sets the iface parameter (implicitly), which is not supported yet
by the libvirt backend.

This updates commit 8e198dc1c50fcdba6720b23da478ef75cc979733.

+++ tests/regressions/rhbz690819.sh
@@ -40,6 +40,12 @@ if [ -n "$SKIP_TEST_RHBZ690819_SH" ]; then
     exit 77
 fi
 
+method="$(../../fish/guestfish get-attach-method)"
+if [[ "$method" =~ ^libvirt ]]; then
+    echo "$0: test skipped because attach-method ($method) is 'libvirt'."
+    exit 77
+fi
+
 rm -f test.img
 
 ../../fish/guestfish sparse test.img 100M
